+++ src/test/scala/grapevineType/TestGrapevineType.scala
@@ -19,7 +19,7 @@ class TestGrapevineType extends FunSuite {
 
   test ("getTypeFromValue test") {
     val values = Array(1, 1.2, null)
-    val types = Array(Some(classOf[IntegerType]), Some(classOf[FloatType]), None)
+    val types = Array(Some(classOf[ByteType]), Some(classOf[FloatType]), None)
 
     values.zipWithIndex.foreach { case(value, index) =>
       val t = types(index)
